fixes #30 pre refactor
