+++ wcag-tool/chrome-extension/src/components/editor/AnnotationListSlider.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react'
+import React from 'react'
 import SlidingPane from 'react-sliding-pane'
 import { usePopperTooltip } from 'react-popper-tooltip'
 import './react-sliding-pane.css'
@@ -11,32 +11,10 @@ import NoAnnotation from './NoAnnotation'
 
 const AnnotationListSlider = () => {
   const { annotations } = useAnnotation()
-  const [{ openElementSelector, closeAllSliders }, { listSliderIsOpen, elementSelectorIsOpen }] =
-    useSliders()
-  const editorIframe = document.querySelector('.editor')
+  const [{ openElementSelector, closeAllSliders }, { listSliderIsOpen }] = useSliders()
 
   const { getArrowProps, getTooltipProps, setTooltipRef, setTriggerRef, visible } =
     usePopperTooltip({ placement: 'bottom' })
-
-  useEffect(() => {
-    if (elementSelectorIsOpen) {
-      editorIframe.contentWindow.postMessage(
-        JSON.stringify({
-          method: 'elementSelectionMode',
-          content: true,
-        }),
-        '*'
-      )
-    } else {
-      editorIframe.contentWindow.postMessage(
-        JSON.stringify({
-          method: 'elementSelectionMode',
-          content: false,
-        }),
-        '*'
-      )
-    }
-  }, [elementSelectorIsOpen])
 
   return (
     <SlidingPane
