+++ wcag-tool/chrome-extension/src/views/editor.jsx
@@ -1,51 +1,14 @@
-import React, { useEffect } from 'react'
+import React from 'react'
 import ReactDOM from 'react-dom'
 import { Provider } from 'react-redux'
 import '../css/styles.css'
 import './editor.css'
-import './editorjs'
 import App from '../components/editor/App'
 import store from '../services/store'
-import { useSliders } from '../hooks'
+import { useRegisterEditorEffects } from '../hooks/editor.hooks'
 
 const Editor = () => {
-  const [{ openCreateAndEditSlider }] = useSliders()
-
-  useEffect(() => {
-    const eventListener = (event) => {
-      const message = JSON.parse(event.data)
-      if (message.method === 'onUpdate') {
-        tabData.docSaved = message.saved
-      }
-      if (message.method === 'onInit') {
-        tabData.options.disableFormatPage = !message.formatPageEnabled
-        // formatPageButton.hidden = !message.formatPageEnabled;
-        document.title = '[WCAG] ' + message.title
-        if (message.filename) {
-          tabData.filename = message.filename
-        }
-        if (message.icon) {
-          const linkElement = document.createElement('link')
-          linkElement.rel = 'icon'
-          linkElement.href = message.icon
-          document.head.appendChild(linkElement)
-        }
-        tabData.docSaved = true
-      }
-
-      if (message.method === 'onElementSelected') {
-        console.log('onElementSelected', message)
-        openCreateAndEditSlider(message.content)
-      }
-    }
-
-    addEventListener('message', eventListener)
-
-    return () => {
-      removeEventListener('message', eventListener)
-    }
-  }, [])
-
+  useRegisterEditorEffects()
   return <App />
 }
